requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torchvision
timm==0.9.7
scipy
matplotlib
tensorboard
open-clip-torch==2.24.0
scikit-learn==1.3.2
wandb
tqdm
fairscale
sentencepiece
Pillow
opencv-python
gradio
autofaiss
tensorboard
diffusers
plotly
easydict
https://github.com/unlimblue/KNN_CUDA/releases/download/0.2/KNN_CUDA-0.2-py3-none-any.whl
huggingface_hub
transformers
open3d
openai==0.28